% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold4GPD.R
\name{Data_Randomization}
\alias{Data_Randomization}
\title{Data_Randomization}
\usage{
Data_Randomization(x, random_digit = 2, round_digit = 4, seed = 1)
}
\arguments{
\item{x}{- data vector}

\item{random_digit}{- the precision of data vector}

\item{round_digit}{- the precision of return randomization data vector}

\item{seed}{- presetting random seed}
}
\value{
RandomData - randomization data vector
}
\description{
add the randomization of data to reduce the limitation of minimum measured value
}
\examples{
x <- rnorm(100)
x <- sort(x)
x <- round(x*10)/10
Data_Randomization(x)

}

