% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold4GPD.R
\name{E.Info.Mat}
\alias{E.Info.Mat}
\title{E.Info.Mat}
\usage{
E.Info.Mat(theta, u, M)
}
\arguments{
\item{theta}{- vector of parameters (mu,sigma, xi)}

\item{u}{- threshold for NHPP}

\item{M}{- number of superpositions or "blocks" / "years" the process corresponds to}
}
\value{
EIM - expected information matrix

Errors - vector of errors from the numerical integration of the 6 unique components
}
\description{
Calculates the numerically-integrated expected information matrix for an NHPP with specified parameters
}

