% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold4GPD.R
\name{Expl.diag}
\alias{Expl.diag}
\title{producing the exponential diagnostic plots}
\usage{
Expl.diag(x, u = NULL, k, q1, q2 = 1, nbs = 1000, alpha = 0.05,
  plots = c("LRT", "WN", "PS"), UseQuantiles = TRUE, param = "InvRate",
  pmar = c(5.5, 7, 3, 3), ...)
}
\arguments{
\item{x}{- data vector}

\item{u}{- optional; vector of candidate thresholds}

\item{k}{- number of thresholds to consider (if u unspecified)}

\item{q1}{-  lowest quantile for the threshold sequence}

\item{q2}{-  upper quantile limit for the threshold sequence (q2 itself is not used as a threshold, but rather the uppermost}

\item{nbs}{- number of simulations used to assess the null distribution of the LRT, and produce the p-value}

\item{alpha}{- significance level of the LRT}

\item{plots}{- which plots to produce; "LRT"= likelihood ratio test, "WN" = white noise, "PS" = parameter stability}

\item{UseQuantiles}{- logical; use quantiles as the thresholds in the plot?}

\item{param}{- character specifying "InvRate" or "Rate" for either inverse rate parameter / rate parameter, respectively}

\item{pmar}{- vector of length 4 giving the arguments for the plot margins in par(mar=c(*,*,*,*))}

\item{...}{- other parameter}
}
\value{
MLE - MLEs from all thresholds

Cov - Joint asymptotic covariance matrix for xi

WN - values of the white noise process

LRT - values of the LT test statistic vs threshold

pval - p-value of the LR test

k - final number of thresholds used

thresh - threshold selected by the LR procedure

mle.u - MLE from selected threshold
}
\description{
Function to produce diagnostic plots and test statistics for the rate / inverse rate parameter of the Exponential model
}
\examples{
library(mvtnorm)
xbvn<-rmvnorm(6000, sigma=matrix(c(1,0.7,0.7,1),2,2))

# Transform margins to exponential
xbvn.exp<- -log(1-pnorm(xbvn))
Expl.diag(apply(xbvn.exp,1,min), k=30, q1=0, param="Rate")
Expl.diag(apply(xbvn.exp,1,min), k=30, q1=0, param="InvRate")

}

