% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold4GPD.R
\name{Joint.MLE.Expl}
\alias{Joint.MLE.Expl}
\title{Joint.MLE.Expl}
\usage{
Joint.MLE.Expl(x, u = NULL, k, q1, q2 = 1, param)
}
\arguments{
\item{x}{- vector of data}

\item{u}{- vector of thresholds. If not supplied, then k thresholds between quantiles (q1, q2) will be used}

\item{k}{- number of thresholds to consider if u not supplied}

\item{q1}{- lowest quantile for the threshold sequence}

\item{q2}{-  upper quantile limit for the threshold sequence (q2 itself is not used as a threshold, but rather the uppermost
threshold will be at the (q2-1/k) quantile)}

\item{param}{- character specifying "InvRate" or "Rate" for either inverse rate parameter / rate parameter, respectively}
}
\value{
mle - vector of MLEs above the supplied thresholds

cov - joint asymptotic covariance matrix of these MLEs
}
\description{
Calculates the MLEs of the rate parameter, and joint asymptotic covariance matrix of these MLEs
over a range of thresholds as supplied by the user.
}

