% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold4GPD.R
\name{Joint.MLE.NHPP}
\alias{Joint.MLE.NHPP}
\title{joint maximum likelihood estimation Non-Homogeneous Poisson Process}
\usage{
Joint.MLE.NHPP(x, u = NULL, k, q1, q2 = 1, par, M, MLE_model = "median")
}
\arguments{
\item{x}{- vector of data}

\item{u}{- vector of thresholds. If not supplied, then k thresholds between quantiles (q1, q2) will be used}

\item{k}{- number of thresholds to consider if u not supplied}

\item{q1}{- lowest quantile for the threshold sequence}

\item{q2}{-  upper quantile limit for the threshold sequence (q2 itself is not used as a threshold, but rather the uppermost
threshold will be at the (q2-1/k) quantile)}

\item{par}{- starting values for the optimization}

\item{M}{- number of superpositions or "blocks" / "years" the process corresponds to (affects estimation of mu, sigma,
but these can be changed post-hoc to correspond to any number)}

\item{MLE_model}{- which model of maximum likelihood estimation is used;
"median"= use the median data of 7 maximum likelihood estimation (not accurate enough but suitable for most data vector)
"min"= use the min data of 7 maximum likelihood estimation (most accurate but not suitable for some small data vector)
"second_min"= use the second min data of 7 maximum likelihood estimation (almost same as "min")}
}
\value{
mle - matrix of MLEs above the supplied thresholds; columns are (mu, sigma, xi)

Cov.all - joint asymptotic covariance matrix of all MLEs

Cov.mu - joint asymptotic covariance matrix of MLEs for mu

Cov.sig - joint asymptotic covariance matrix of MLEs for sig

Cov.xi - joint asymptotic covariance matrix of MLEs for xi
}
\description{
the MLEs of the parameters (mu,sigma,xi), and joint asymptotic covariance matrix of these MLEs
over a range of thresholds as supplied by the user.
}

