% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold4GPD.R
\name{K.ForwardStop}
\alias{K.ForwardStop}
\title{K.ForwardStop}
\usage{
K.ForwardStop(pvalues, alpha = 0.05)
}
\arguments{
\item{pvalues}{- the p-values for each threshold in the threshold list}

\item{alpha}{- the pre-specified significance level, usually take 0.01 or 0.05}
}
\value{
forward - if the given null hypothesis should be rejected. Take 0 as no rejection and 1 as make rejection
}
\description{
calculate the cutoff k based on ForwardStop
}

