% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold4GPD.R
\name{KvsN}
\alias{KvsN}
\title{KvsN}
\usage{
KvsN(x, wetCuttingPoint = 0.1)
}
\arguments{
\item{x}{- data vector}

\item{wetCuttingPoint}{- the minimum precision which can be considered as "rain"}
}
\value{
ks - an appropriate number of thresholds
}
\description{
Function to determine an appropriate number of thresholds
}

