% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold4GPD.R
\name{NHPP.diag}
\alias{NHPP.diag}
\title{producing the Non-Homogeneous Poisson Process diagnostic plots}
\usage{
NHPP.diag(x, u = NULL, k, q1 = NULL, q2 = 1, par = NULL, M = NULL,
  nbs = 1000, alpha = 0.05, plots = c("LRT", "WN", "PS", "LRTEST", "KS",
  "AD"), UseQuantiles = TRUE, DEBUGing = FALSE, pmar = c(5.5, 7, 3, 3),
  head = 5, MLE_model = NULL, use_Data_Randomization = FALSE,
  use_Threshold_Randomization = FALSE, ...)
}
\arguments{
\item{x}{- data vector}

\item{u}{- optional; vector of candidate thresholds}

\item{k}{- number of thresholds to consider (if u unspecified)}

\item{q1}{-  lowest quantile for the threshold sequence}

\item{q2}{-  upper quantile limit for the threshold sequence (q2 itself is not used as a threshold, but rather the uppermost
threshold will be at the (q2-1/k) quantile)}

\item{par}{- starting values for the optimization}

\item{M}{- number of superpositions or "blocks" / "years" the process corresponds to (can affect the optimization)}

\item{nbs}{- number of simulations used to assess the null distribution of the LRT, and produce the p-value}

\item{alpha}{- significance level of the LRT}

\item{plots}{- which plots to produce;
"LRT"= likelihood ratio
"WN" = white noise
"PS" = parameter stability
"LRTEST" = likelihood ratio test for assistance
"KS" = Kolmogorov-Smirnov goodness-of-fit test for assistance
"AD" = Anderson-Darling test for assistance}

\item{UseQuantiles}{- logical; use quantiles as the thresholds in the plot?}

\item{DEBUGing}{- default as false, set to true if you want to debug}

\item{pmar}{- vector of length 4 giving the arguments for the plot margins in par(mar=c(*,*,*,*))}

\item{head}{- number of thresholds inserted between the first threshold and the second threshold temporary to reduce the
variance at the begining of likelihood ratio statistic test}

\item{MLE_model}{- which model of maximum likelihood estimation is used;
"median"= use the median data of 7 maximum likelihood estimation (not accurate enough but suitable for most data vector)
"min"= use the min data of 7 maximum likelihood estimation (most accurate but not suitable for some small data vector)
"second_min"= use the second min data of 7 maximum likelihood estimation (almost same as "min")}

\item{use_Data_Randomization}{- whether or not use Data_Randomization to randomize the given data}

\item{use_Threshold_Randomization}{- whether or not use Threshold_Randomization to randomize the thresholds}

\item{...}{- other parameter}
}
\value{
MLEall - MLEs from all thresholds

Cov.xi - Joint asymptotic covariance matrix for xi

WN - values of the white noise process

LRT - values of the LT test statistic vs threshold

pval - p-value of the LR test

k - final number of thresholds used

thresh - threshold selected by LR procedure

mle.u - MLE from selected threshold
}
\description{
Function to produce diagnostic plots and test statistics for the NHPP model
}
\details{
Function to produce diagnostic plots and test statistics for the NHPP model
}
\examples{
## insert an easy example for test run
set.seed(1)
xnorm<-abs(rnorm(5000))
thresholds_xnorm <- Threshold_Generator_Uniform(xnorm, wetCuttingPoint=0.1)
ChangedThresholds_xnorm <- Threshold_Randomization(xnorm,thresholds_xnorm,seed = 1)
nhpp <-NHPP.diag(xnorm, u= ChangedThresholds_xnorm, plots=c("LRT", "WN", "PS"))

##insert an example for test run
set.seed(1)
x <- rnorm(100000)
x <- x[x > quantile(x, probs = 0.9)]
x <- sort(x)
x <- Data_Randomization(x,seed = 1)
x <- sort(x)
thresholds_x <- Threshold_Generator_Uniform(x, wetCuttingPoint=0.1)
ChangedThresholds_x <- Threshold_Randomization(x,thresholds_x,seed = 1)
nhpp <-NHPP.diag(x, u= ChangedThresholds_x,
                 M=365, nbs=1000, alpha=0.05, plots=c("LRT", "WN", "PS"), UseQuantiles=FALSE,
                 cex.lab=1.5, cex.axis=1.4,cex.main=2,mgp=c(4.2,1,0))

#########################
# View different plots for easy example:
# likelihood ratio plot only
NHPP.diag(xnorm, u= ChangedThresholds_xnorm, plots=c("LRT"))

# white noise plot only
NHPP.diag(xnorm, u= ChangedThresholds_xnorm, plots=c("WN"))

# parameter stability plot only
nhpp <-NHPP.diag(xnorm, u= ChangedThresholds_xnorm, plots=c("PS"))

# likelihood ratio test for assistance plot only
NHPP.diag(xnorm, u= ChangedThresholds_xnorm, plots=c("LRTEST"))

# Kolmogorov-Smirnov goodness-of-fit test for assistance plot only
NHPP.diag(xnorm, u= ChangedThresholds_xnorm, plots=c("KS"))

# Anderson-Darling test for assistance plot only
NHPP.diag(xnorm, u= ChangedThresholds_xnorm, plots=c("AD"))

# a showoff of all the plots
NHPP.diag(xnorm, u= ChangedThresholds_xnorm, plots=c("LRT", "WN", "PS", "LRTEST", "KS", "AD"))

}

