% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold4GPD.R
\name{nhpp.nll}
\alias{nhpp.nll}
\title{Non-Homogeneous Poisson Process negative log-likelihood}
\usage{
nhpp.nll(theta, x, u, M)
}
\arguments{
\item{theta}{- parameter vector (mu, sigma, xi)}

\item{x}{- data vector}

\item{u}{- threshold}

\item{M}{- number of superpositions or "blocks" / "years" the process corresponds to (affects estimation of mu, sigma,
but these can be changed post-hoc to correspond to any number)}
}
\value{
negative log-likelihood value
}
\description{
Negative log likelihood for the NHPP model, to minimize for MLEs
}

