% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold4GPD.R
\name{norm.LRT}
\alias{norm.LRT}
\title{normal distribution likelihood ratio statistics}
\usage{
norm.LRT(x, u)
}
\arguments{
\item{x}{- vector of white noise process (WNP, usually normalized estimates of xi or the exponential rate paramter 1/eta)}

\item{u}{- vector of thresholds that are associated to the WNP}
}
\value{
v

likelihood ratio
}
\description{
Evaluates the likelihood ratio statistics for testing white noise
}

