% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold4GPD.R
\name{Threshold_Generator_Density}
\alias{Threshold_Generator_Density}
\title{Thershold_Generator_Density}
\usage{
Threshold_Generator_Density(x, wetCuttingPoint = 0.1, ks = NULL, q = NULL)
}
\arguments{
\item{x}{- data vector}

\item{wetCuttingPoint}{- the minimum precision which can be considered as "rain"}

\item{ks}{- number of candidate thresholds}

\item{q}{- presetting value to control the density (usually set between 1 and 2)
(the bigger q is, the more intensive thresholds in tail part)}
}
\value{
thresholds - candidate thresholds list
}
\description{
generate a candidate thresholds list for given data set based on variant distance among the distribution density
}
\examples{
set.seed(1)
xnorm<-abs(rnorm(5000))
Threshold_Generator_Density(xnorm)

}

