% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold4GPD.R
\name{Threshold_Generator_Uniform}
\alias{Threshold_Generator_Uniform}
\title{Thershold_Generator_Uniform}
\usage{
Threshold_Generator_Uniform(x, wetCuttingPoint = 0.1, ks = NULL)
}
\arguments{
\item{x}{- data vector}

\item{wetCuttingPoint}{- the minimum precision which can be considered as "rain"}

\item{ks}{- number of candidate thresholds}
}
\value{
thresholds - candidate thresholds list
}
\description{
generate a candidate thresholds list for given data set based on uniform distance among the distribution
}
\examples{
set.seed(1)
xnorm<-abs(rnorm(5000))
Threshold_Generator_Uniform(xnorm)

}

