% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold4GPD.R
\name{Threshold_Randomization}
\alias{Threshold_Randomization}
\title{Threshold_Randomization}
\usage{
Threshold_Randomization(x, thresholds, seed = 1)
}
\arguments{
\item{x}{- data vector}

\item{thresholds}{- thresholds vector}

\item{seed}{- presetting random seed}
}
\value{
thresholds - randomization thresholds vector
}
\description{
add the randomization of thresholds to increase the randomization of thresholds
}
\examples{
set.seed(1)
xnorm<-abs(rnorm(5000))
thresholds_xnorm <- Threshold_Generator_Uniform(xnorm)
thresholds_xnorm
Threshold_Randomization(xnorm,thresholds_xnorm,seed = 1)

}

